Moved PssClient into Pss namespace
