+++ tests/PssClientTest.php
@@ -7,7 +7,7 @@ use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\TestCase;
-use UKFast\PssClient;
+use UKFast\Pss\PssClient;
 
 class PssClientTest extends TestCase
 {
